fix: rename burger comstructors component
