+++ src/features/burger-constructor/componets/index.ts
@@ -1,2 +1,2 @@
+export * from './empty-constructor'
 export * from './empty-item'
-export * from './empty-burger-constructor'
